Move `DisplayFolder` related implementation to from `FolderRepository` to `DisplayFolderRepository`
