+++ legacy/core/src/main/java/com/fsck/k9/mailstore/KoinModule.kt
@@ -1,5 +1,6 @@
 package com.fsck.k9.mailstore
 
+import app.k9mail.legacy.mailstore.DisplayFolderRepository
 import app.k9mail.legacy.mailstore.FolderRepository
 import app.k9mail.legacy.mailstore.MessageListRepository
 import app.k9mail.legacy.mailstore.MessageStoreManager
@@ -9,7 +10,18 @@ import com.fsck.k9.message.extractors.MessagePreviewCreator
 import org.koin.dsl.module
 
 val mailStoreModule = module {
-    single { FolderRepository(messageStoreManager = get(), accountManager = get()) }
+    single {
+        FolderRepository(
+            messageStoreManager = get(),
+            accountManager = get(),
+        )
+    }
+    single {
+        DisplayFolderRepository(
+            messageStoreManager = get(),
+            accountManager = get(),
+        )
+    }
     single { MessageViewInfoExtractorFactory(get(), get(), get()) }
     single { StorageManager.getInstance(get()) }
     single { SpecialFolderSelectionStrategy() }
