+++ legacy/mailstore/src/main/java/app/k9mail/legacy/mailstore/RemoteFolderTypeMapper.kt
@@ -1,0 +1,18 @@
+package app.k9mail.legacy.mailstore
+
+import app.k9mail.core.mail.folder.api.FolderType
+import com.fsck.k9.mail.FolderType as RemoteFolderType
+
+object RemoteFolderTypeMapper {
+
+    fun RemoteFolderType.toFolderType(): FolderType = when (this) {
+        RemoteFolderType.REGULAR -> FolderType.REGULAR
+        RemoteFolderType.INBOX -> FolderType.INBOX
+        RemoteFolderType.OUTBOX -> FolderType.REGULAR // We currently don't support remote Outbox folders
+        RemoteFolderType.DRAFTS -> FolderType.DRAFTS
+        RemoteFolderType.SENT -> FolderType.SENT
+        RemoteFolderType.TRASH -> FolderType.TRASH
+        RemoteFolderType.SPAM -> FolderType.SPAM
+        RemoteFolderType.ARCHIVE -> FolderType.ARCHIVE
+    }
+}
