+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/managefolders/ManageFoldersViewModel.kt
@@ -5,9 +5,11 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.folder.DisplayFolder
-import app.k9mail.legacy.mailstore.FolderRepository
+import app.k9mail.legacy.mailstore.DisplayFolderRepository
 
-class ManageFoldersViewModel(private val folderRepository: FolderRepository) : ViewModel() {
+class ManageFoldersViewModel(
+    private val folderRepository: DisplayFolderRepository,
+) : ViewModel() {
     fun getFolders(account: Account): LiveData<List<DisplayFolder>> {
         return folderRepository.getDisplayFoldersFlow(account.uuid).asLiveData()
     }
